packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shout-run-sdk"
version = "0.1.0"
description = "Python SDK for shout.run — programmatic terminal broadcasting"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pavan Madiraju" },
]
keywords = ["terminal", "broadcast", "live", "sdk", "agent", "ai", "streaming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
dependencies = [
    "websocket-client>=1.6.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://shout.run"
Repository = "https://github.com/pavanmadiraju91/shout-run"
Documentation = "https://github.com/pavanmadiraju91/shout-run/tree/main/packages/sdk-python"

[tool.hatch.build.targets.wheel]
packages = ["src/shout_sdk"]
